requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai-whisper==20240930
youtube-transcript-api==0.6.2
pydantic==2.9.2
httpx==0.27.2
